Catálogo de las series de valores gubernamentales en México.

Principales series de tiempo relacionadas a los valores gubernamentales en México para consulta en la API del banco nacional.

Usage

catalogo_series

Format

A data frame with 16 rows and 4 variables:

Instrumento

Nombre del instrumento financiero

Plazo

Número de días o años asignados en la subasta semanal

Monto_Asignado

Cantidad de dinero asignada por el banco nacional a este intrumento

Tasa_Rendimiento

Tasa de rendimiento derivada de la subasta semanal

Source

Banxico